Fix create_file_writer() memory leak in graph mode
This fixes #26684, a memory leak that occurred when repeatedly calling create_file_writer() for new graphs in either tf.contrib.summary or TF 2.0's tf.summary.  The implementation for tf.contrib.summary.summary_writer_initializer_op() held strong references to all graphs that had ever created summary writers, even when those graphs were otherwise unreachable.

We address this by reverting the implementation to use graph collections, since while deprecated, the API this is needed for is only exposed in TF 1.x.

I added a test and confirmed that it failed prior to the fix, and passes after it.

PiperOrigin-RevId: 241174020