Make CustomKernelCreator into a abstract class
Instead of a function, we make CustomKernelCreator into an abstract
class in order to add a CanCreateKernel method. This allows us to
distinguish between unsupported operations and an actual failure when
creating a kernel. As a result, we also rename CreateXlaLaunchOp to
XlaKernelCreator.

We also remove the CustomCreatorSingleton since only eager requires the
custom kernel creator to be set. As a result, we are able to further
simply the flr construction signature.

Note that noinline tests needed to be disabled since this change
showed the tests were not operating as intended.

PiperOrigin-RevId: 241625510