Remove checking if argument Tensor is initialized in call frame.
In the existing code, when an argument Tensor is not initialized, it falls through the if-else chain and the first else branch catches it, and will reference out-of-range memory region in vector.

Looking at the code, there does not seem to be a strong reason to check whether an argument is initialized or not. So removing that condition in if cluase.

PiperOrigin-RevId: 242015796